+++ crates/release_plz/src/args/update.rs
@@ -119,6 +119,7 @@ impl Update {
         cargo_metadata: cargo_metadata::Metadata,
     ) -> anyhow::Result<UpdateRequest> {
         let project_manifest = self.project_manifest();
+        check_if_cargo_lock_is_ignored(&project_manifest)?;
         let mut update = UpdateRequest::new(cargo_metadata)
             .with_context(|| {
                 format!("Cannot find file {project_manifest:?}. Make sure you are inside a rust project or that --project-manifest points to a valid Cargo.toml file.")
@@ -160,6 +161,9 @@ impl Update {
         }
         if let Some(registry) = &self.registry {
             update = update.with_registry(registry.clone());
+        }
+        if let Some(release_commits) = config.workspace.release_commits {
+            update = update.with_release_commits(release_commits.clone())?;
         }
 
         Ok(update)
@@ -203,6 +207,17 @@ impl Update {
             .as_deref()
             .or(config.workspace.changelog_config.as_deref())
     }
+}
+
+fn check_if_cargo_lock_is_ignored(local_manifest: &Path) -> anyhow::Result<()> {
+    let repo_path = release_plz_core::root_repo_path(local_manifest)?;
+    let cargo_lock_path = local_manifest.with_file_name("Cargo.lock");
+    let is_cargo_lock_ignored = git_cmd::is_file_ignored(&repo_path, &cargo_lock_path)?;
+    anyhow::ensure!(
+        !(is_cargo_lock_ignored && cargo_lock_path.exists()),
+        "Cargo.lock is present in your .gitignore and is also committed. Remove it from your repository or from your `.gitignore` file."
+    );
+    Ok(())
 }
 
 #[cfg(test)]