+++ website/sidebars.js
@@ -34,7 +34,7 @@ const sidebars = {
       label: "GitHub Action",
       collapsed: false,
       link: { type: "doc", id: "github/index" },
-      items: ["github/quickstart", "github/token", "github/update"],
+      items: ["github/quickstart", "github/token", "github/update", "github/advanced"],
     },
     "config",
     "changelog-format",
@@ -47,7 +47,7 @@ const sidebars = {
       label: "Extra",
       collapsed: false,
       link: { type: "doc", id: "extra/index" },
-      items: ["extra/releasing-binaries", "extra/single-changelog", "extra/yanked-packages"],
+      items: ["extra/releasing-binaries", "extra/single-changelog", "extra/single-tag", "extra/yanked-packages"],
     },
   ],
 };